Refactor compilation of shaders in render pass
Instead of re-compiling the same hard-coded shaders every frame, they are compiled once and passed to the start of a render pass.

* use `include_str!` macro to load shaders from files
